server/requirements.txt
bcrypt==4.1.3
blinker==1.8.2
boto3==1.34.133
botocore==1.34.133
click==8.1.7
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
s3transfer==0.10.2
six==1.16.0
urllib3==2.2.2
Werkzeug==3.0.3
